# app/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScoutEvent {
  id: string;
  name: string;
  nameEn?: string;
  unit: string;
  date: string;
  helpers: number;
}

export interface LeaderPreference {
  id: string;
  createdAt: string;
  fullName: string;
  gender: string;
  unit: string;
  yearsExp: number;
  isSenior: boolean;
  targetIcCount: number;
  skills: string[];
  preferredIcEvents: string[];
  helperEvents: string[];
  preferredPartners: string[];
  notes: string;
}

export type LeaderPreferenceInputGender = typeof LeaderPreferenceInputGender[keyof typeof LeaderPreferenceInputGender];


export const LeaderPreferenceInputGender = {
  男: '男',
  女: '女',
} as const;

export interface LeaderPreferenceInput {
  /** @minLength 1 */
  fullName: string;
  gender: LeaderPreferenceInputGender;
  unit: string;
  /** @minimum 0 */
  yearsExp: number;
  isSenior: boolean;
  /**
     * @minimum 2
     * @maximum 4
     */
  targetIcCount: number;
  /** @maxItems 5 */
  skills: string[];
  preferredIcEvents: string[];
  helperEvents: string[];
  preferredPartners: string[];
  notes: string;
}

export type ScoutSummaryTopSkillsItem = {
  skill: string;
  count: number;
};

export interface ScoutSummary {
  totalSubmissions: number;
  seniorCount: number;
  topSkills: ScoutSummaryTopSkillsItem[];
}

export interface ScoutChatInput {
  /**
     * @minLength 1
     * @maxLength 800
     */
  message: string;
}

export interface ScoutChatReply {
  reply: string;
}