# app/lib/api-zod/src/generated/types/leaderPreference.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderPreference {
  id: string;
  createdAt: string;
  fullName: string;
  gender: string;
  unit: string;
  yearsExp: number;
  isSenior: boolean;
  targetIcCount: number;
  skills: string[];
  preferredIcEvents: string[];
  helperEvents: string[];
  preferredPartners: string[];
  notes: string;
}