# app/lib/api-zod/src/generated/types/leaderPreferenceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderPreferenceInputGender } from './leaderPreferenceInputGender';

export interface LeaderPreferenceInput {
  /** @minLength 1 */
  fullName: string;
  gender: LeaderPreferenceInputGender;
  unit: string;
  /** @minimum 0 */
  yearsExp: number;
  isSenior: boolean;
  /**
     * @minimum 2
     * @maximum 4
     */
  targetIcCount: number;
  /** @maxItems 5 */
  skills: string[];
  preferredIcEvents: string[];
  helperEvents: string[];
  preferredPartners: string[];
  notes: string;
}