# app/lib/api-zod/src/generated/types/scoutChatInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoutChatInput {
  /**
     * @minLength 1
     * @maxLength 800
     */
  message: string;
}