# app/lib/api-zod/src/generated/types/scoutEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoutEvent {
  id: string;
  name: string;
  nameEn?: string;
  unit: string;
  date: string;
  helpers: number;
}