# app/lib/api-zod/src/generated/types/scoutSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScoutSummaryTopSkillsItem } from './scoutSummaryTopSkillsItem';

export interface ScoutSummary {
  totalSubmissions: number;
  seniorCount: number;
  topSkills: ScoutSummaryTopSkillsItem[];
}